Deleted unnecessary annotations, fixed spaces, fixed save method in BookServiceImpl and delete method in CategoryServiceImp

+++ src/main/java/online/book/store/service/category/CategoryServiceImp.java
@@ -51,6 +51,10 @@ public class CategoryServiceImp implements CategoryService {
 
     @Override
     public void deleteById(Long id) {
-        categoryRepository.deleteById(id);
+        if (categoryRepository.existsById(id)) {
+            categoryRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Category with id " + id + " does not exist.");
+        }
     }
 }
